Redémarrage gracieux du process master de gunicorn
`pc -C gunicorn` only lists the processes with `gunicorn` command, i.e., workers and master process. Workers are children of master as can be seen using `ps -C gunicorn fc -o ppid,pid,cmd`. We only need the pid of the master, therefore h flag is used to remove the first line which is PID text. Note that, f flag assures that master is printed above workers.

The correct procedure is to send `HUP` signal only to the master. In this way `gunicorn` is gracefully restarted, only the workers, not master, are recreated.